Eliminate warnings AGG (64 Bits) + Update jpeglibrary. +  Update freetype library.

+++ agg/font_freetype/agg_font_freetype.cpp
@@ -823,7 +823,7 @@ namespace agg
     {
         if(m_cur_face && m_name)
         {
-            unsigned name_len = strlen(m_name);
+            unsigned name_len = (unsigned int)strlen(m_name);
             if(name_len > m_name_len)
             {
                 delete [] m_signature;
